Abstracted most inspiration engine tile functionality
