+++ src/main/java/com/verdantartifice/thaumicwonders/common/tiles/devices/TileMadnessEngine.java
@@ -1,7 +1,23 @@
 package com.verdantartifice.thaumicwonders.common.tiles.devices;
 
-import com.verdantartifice.thaumicwonders.common.tiles.base.TileTW;
+import thaumcraft.api.aspects.Aspect;
 
-public class TileMadnessEngine extends TileTW {
+public class TileMadnessEngine extends AbstractTileResearchEngine {
+    private static final int COST = 10;
+    private static final int CAPACITY = 50;
 
+    @Override
+    public int getCost() {
+        return COST;
+    }
+    
+    @Override
+    public int getCapacity() {
+        return CAPACITY;
+    }
+
+    @Override
+    public Aspect getAspect() {
+        return Aspect.ELDRITCH;
+    }
 }
